requirements.txt
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
conda==4.3.16
DateTime==4.2
google==2.0.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
python-telegram-bot==10.1.0
requests==2.18.4
tables==3.4.3
telepot==12.7
urllib3==1.22
virtualenv==16.0.0
w3lib==1.19.0
